+++ tests/sats.test.ts
@@ -1,7 +1,7 @@
 import { buildApiServer } from '../src/api/init';
 import { cycleMigrations } from '../src/pg/migrations';
 import { PgStore } from '../src/pg/pg-store';
-import { TestFastifyServer } from './helpers';
+import { TestChainhookPayloadBuilder, TestFastifyServer } from './helpers';
 
 describe('/sats', () => {
   let db: PgStore;
@@ -39,31 +39,28 @@ describe('/sats', () => {
   });
 
   test('returns sat with inscription', async () => {
-    await db.insertInscriptionGenesis({
-      inscription: {
-        genesis_id: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dci0',
-        mime_type: 'image/png',
-        content_type: 'image/png',
-        content_length: 5,
-        number: 7,
-        content: '0x48656C6C6F',
-        fee: '2805',
-      },
-      location: {
-        genesis_id: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dci0',
-        block_height: 775617,
-        block_hash: '00000000000000000002a90330a99f67e3f01eb2ce070b45930581e82fb7a91d',
-        tx_id: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dc',
-        address: 'bc1p3cyx5e2hgh53w7kpxcvm8s4kkega9gv5wfw7c4qxsvxl0u8x834qf0u2td',
-        output: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dc:0',
-        offset: '0',
-        value: '10000',
-        timestamp: 1676913207,
-        sat_ordinal: '257418248345364',
-        sat_rarity: 'common',
-        sat_coinbase_height: 650000,
-      },
-    });
+    await db.updateInscriptions(
+      new TestChainhookPayloadBuilder()
+        .apply()
+        .block({ height: 775617 })
+        .transaction({ hash: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dc' })
+        .inscriptionRevealed({
+          content_bytes: '0x48656C6C6F',
+          content_type: 'image/png',
+          content_length: 5,
+          inscription_number: 7,
+          inscription_fee: 2805,
+          inscription_id: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dci0',
+          inscription_output_value: 10000,
+          inscriber_address: 'bc1p3cyx5e2hgh53w7kpxcvm8s4kkega9gv5wfw7c4qxsvxl0u8x834qf0u2td',
+          ordinal_number: 257418248345364,
+          ordinal_block_height: 650000,
+          ordinal_offset: 0,
+          satpoint_post_inscription:
+            '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dc:0:0',
+        })
+        .build()
+    );
     const response = await fastify.inject({
       method: 'GET',
       url: '/ordinals/v1/sats/257418248345364',