+++ tests/server.test.ts
@@ -3,7 +3,7 @@ import { buildChainhookServer, CHAINHOOK_BASE_PATH, PREDICATE_UUID } from '../sr
 import { ENV } from '../src/env';
 import { cycleMigrations } from '../src/pg/migrations';
 import { PgStore } from '../src/pg/pg-store';
-import { TestFastifyServer } from './helpers';
+import { TestChainhookPayloadBuilder, TestFastifyServer } from './helpers';
 
 describe('EventServer', () => {
   let db: PgStore;
@@ -168,15 +168,17 @@ describe('EventServer', () => {
       });
       expect(response.statusCode).toBe(200);
 
-      const query = await db.getInscriptions({
-        genesis_id: ['0268dd9743c862d80ab02cb1d0228036cfe172522850eb96be60cfee14b31fb8i0'],
-        limit: 1,
-        offset: 0,
-      });
+      const query = await db.getInscriptions(
+        {
+          limit: 1,
+          offset: 0,
+        },
+        { genesis_id: ['0268dd9743c862d80ab02cb1d0228036cfe172522850eb96be60cfee14b31fb8i0'] }
+      );
       const inscr = query.results[0];
       expect(inscr).not.toBeUndefined();
       expect(inscr.address).toBe('bc1p3cyx5e2hgh53w7kpxcvm8s4kkega9gv5wfw7c4qxsvxl0u8x834qf0u2td');
-      expect(inscr.content_length).toBe(12);
+      expect(inscr.content_length).toBe('12');
       expect(inscr.content_type).toBe('text/plain;charset=utf-8');
       expect(inscr.genesis_address).toBe(
         'bc1p3cyx5e2hgh53w7kpxcvm8s4kkega9gv5wfw7c4qxsvxl0u8x834qf0u2td'
@@ -184,7 +186,7 @@ describe('EventServer', () => {
       expect(inscr.genesis_block_hash).toBe(
         '163de66dc9c0949905bfe8e148bde04600223cf88d19f26fdbeba1d6e6fa0f88'
       );
-      expect(inscr.genesis_block_height).toBe(107);
+      expect(inscr.genesis_block_height).toBe('107');
       expect(inscr.genesis_fee).toBe('3425');
       expect(inscr.genesis_id).toBe(
         '0268dd9743c862d80ab02cb1d0228036cfe172522850eb96be60cfee14b31fb8i0'
@@ -194,12 +196,12 @@ describe('EventServer', () => {
         '0268dd9743c862d80ab02cb1d0228036cfe172522850eb96be60cfee14b31fb8'
       );
       expect(inscr.mime_type).toBe('text/plain');
-      expect(inscr.number).toBe(100);
+      expect(inscr.number).toBe('100');
       expect(inscr.offset).toBe('0');
       expect(inscr.output).toBe(
         '0268dd9743c862d80ab02cb1d0228036cfe172522850eb96be60cfee14b31fb8:0'
       );
-      expect(inscr.sat_coinbase_height).toBe(25069);
+      expect(inscr.sat_coinbase_height).toBe('25069');
       expect(inscr.sat_ordinal).toBe('125348773618236');
       expect(inscr.sat_rarity).toBe('common');
       expect(inscr.timestamp.toISOString()).toBe('2023-03-03T00:31:50.000Z');
@@ -231,31 +233,34 @@ describe('EventServer', () => {
     });
 
     test('parses inscription_transferred apply and rollback', async () => {
-      await db.insertInscriptionGenesis({
-        inscription: {
-          genesis_id: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dci0',
-          mime_type: 'image/png',
-          content_type: 'image/png',
-          content_length: 5,
-          number: 7,
-          content: '0x48656C6C6F',
-          fee: '2805',
-        },
-        location: {
-          genesis_id: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dci0',
-          block_height: 775617,
-          block_hash: '00000000000000000002a90330a99f67e3f01eb2ce070b45930581e82fb7a91d',
-          tx_id: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dc',
-          address: 'bc1p3cyx5e2hgh53w7kpxcvm8s4kkega9gv5wfw7c4qxsvxl0u8x834qf0u2td',
-          output: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dc:0',
-          offset: '0',
-          value: '10000',
-          timestamp: 1676913207,
-          sat_ordinal: '5',
-          sat_rarity: 'common',
-          sat_coinbase_height: 0,
-        },
-      });
+      await db.updateInscriptions(
+        new TestChainhookPayloadBuilder()
+          .apply()
+          .block({
+            height: 775617,
+            hash: '0x00000000000000000002a90330a99f67e3f01eb2ce070b45930581e82fb7a91d',
+            timestamp: 1676913207,
+          })
+          .transaction({
+            hash: '0x38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dc',
+          })
+          .inscriptionRevealed({
+            content_bytes: '0x48656C6C6F',
+            content_type: 'image/png',
+            content_length: 5,
+            inscription_number: 7,
+            inscription_fee: 2805,
+            inscription_id: '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dci0',
+            inscription_output_value: 10000,
+            inscriber_address: 'bc1p3cyx5e2hgh53w7kpxcvm8s4kkega9gv5wfw7c4qxsvxl0u8x834qf0u2td',
+            ordinal_number: 5,
+            ordinal_block_height: 0,
+            ordinal_offset: 0,
+            satpoint_post_inscription:
+              '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dc:0:0',
+          })
+          .build()
+      );
       const transfer = {
         block_identifier: {
           index: 775618,
@@ -317,15 +322,17 @@ describe('EventServer', () => {
       });
       expect(response.statusCode).toBe(200);
 
-      const query = await db.getInscriptions({
-        genesis_id: ['38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dci0'],
-        limit: 1,
-        offset: 0,
-      });
+      const query = await db.getInscriptions(
+        {
+          limit: 1,
+          offset: 0,
+        },
+        { genesis_id: ['38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dci0'] }
+      );
       const inscr = query.results[0];
       expect(inscr).not.toBeUndefined();
       expect(inscr.address).toBe('bc1p3cyx5e2hgh53w7kpxcvm8s4kkega9gv5wfw7c4qxsvxl0u8x834qf00000');
-      expect(inscr.content_length).toBe(5);
+      expect(inscr.content_length).toBe('5');
       expect(inscr.content_type).toBe('image/png');
       expect(inscr.genesis_address).toBe(
         'bc1p3cyx5e2hgh53w7kpxcvm8s4kkega9gv5wfw7c4qxsvxl0u8x834qf0u2td'
@@ -333,7 +340,7 @@ describe('EventServer', () => {
       expect(inscr.genesis_block_hash).toBe(
         '00000000000000000002a90330a99f67e3f01eb2ce070b45930581e82fb7a91d'
       );
-      expect(inscr.genesis_block_height).toBe(775617);
+      expect(inscr.genesis_block_height).toBe('775617');
       expect(inscr.genesis_fee).toBe('2805');
       expect(inscr.genesis_id).toBe(
         '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dci0'
@@ -343,12 +350,12 @@ describe('EventServer', () => {
         '38c46a8bf7ec90bc7f6b797e7dc84baa97f4e5fd4286b92fe1b50176d03b18dc'
       );
       expect(inscr.mime_type).toBe('image/png');
-      expect(inscr.number).toBe(7);
+      expect(inscr.number).toBe('7');
       expect(inscr.offset).toBe('5000');
       expect(inscr.output).toBe(
         '0268dd9743c862d80ab02cb1d0228036cfe172522850eb96be60cfee14b31fb8:0'
       );
-      expect(inscr.sat_coinbase_height).toBe(0);
+      expect(inscr.sat_coinbase_height).toBe('0');
       expect(inscr.sat_ordinal).toBe('5');
       expect(inscr.sat_rarity).toBe('common');
       expect(inscr.timestamp.toISOString()).toBe('2023-03-03T00:31:50.000Z');